fixing a minor bug in mcp4352 driver

+++ src/Zetaohm_MCP4352/Zetaohm_MCP4352.h
@@ -15,7 +15,7 @@ public:
 	Zetaohm_MCP4352();
   void initialize(uint8_t csPin);
   void setResistance(uint8_t port, uint8_t value);
-  uint8_t readResistance(uint8_t port);
+  uint16_t readResistance(uint8_t port);
   uint16_t 	readAddress(byte addr);
   void			writeByte(byte addr, byte data);
